# backend/internal/httpapi/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API do LogN. Erro de rota que o app lê sai como {\"code\", \"message\"}, e o código é contrato.",
        "title": "LogN API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/internal/licenses/appeal": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Só com ID token OIDC do Google, da conta `ADMIN_SERVICE_ACCOUNT` (ADR 0021). Campo desconhecido no corpo é 400.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Responde à contestação de uma revogação",
                "parameters": [
                    {
                        "description": "Conta, trilha e desfecho",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.licenseActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email_sent": {
                                    "type": "boolean"
                                },
                                "locale": {
                                    "type": "string"
                                },
                                "outcome": {
                                    "type": "string"
                                },
                                "reason": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid_request",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "license_not_revoked, license_appeal_out_of_order, license_store_revoked",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "429": {
                        "description": "rate_limited",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "503": {
                        "description": "internal (sem mailer)",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    }
                }
            }
        },
        "/api/v1/internal/licenses/revoke": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Só com ID token OIDC do Google, da conta `ADMIN_SERVICE_ACCOUNT` (ADR 0021). Campo desconhecido no corpo é 400.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Revoga uma licença à mão",
                "parameters": [
                    {
                        "description": "Conta, trilha e motivo",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.licenseActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email_sent": {
                                    "type": "boolean"
                                },
                                "locale": {
                                    "type": "string"
                                },
                                "outcome": {
                                    "type": "string"
                                },
                                "reason": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid_request",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "license_not_active, license_store_revoked",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "429": {
                        "description": "rate_limited",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "503": {
                        "description": "internal (sem mailer)",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    }
                }
            }
        },
        "/api/v1/internal/play/voided": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Só com ID token OIDC do Google, da conta do Scheduler (ADR 0022). Erros de autorização e de loja saem em texto.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Reembolsos e estornos do Google Play",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "acknowledged": {
                                    "type": "integer"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "voided": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "rate_limited",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Google Play disabled",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/internal/purge": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Só com ID token OIDC do Google, da conta `CLOUD_SCHEDULER_SERVICE_ACCOUNT`. Erros saem em texto, não em `{code}`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Purga diária",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "login_attempts_purged": {
                                    "type": "integer"
                                },
                                "purged": {
                                    "type": "integer"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "waitlist_purged": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpapi.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "httpapi.licenseActionRequest": {
            "type": "object",
            "properties": {
                "evidence": {
                    "type": "string"
                },
                "outcome": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "track_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token do login, no formato `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}